Everything
Contains all code regarding the connect 4 bot as of 1/26/18. This
includes training and mating algorithms, as well as strat4 and strat42,
which are the convergent 'solution' that I have found. This also
includes the ability to play against any AI generated, as well as
nueralView, a tool to help see what patterns the number generate, and
develop an intuition for how the net works.